Create project file structure & Setup basic API
Created the application file structure according to the MVC pattern.
Setup a bare-bones implementation of the express backend API with only
the home (root) route ('/') implented. Installed the pug template engine
for dynamically rendered HTML which can be served by the API. Figured
this is a solid addition since the HTML will be updating real-time.