Implement socket.io connection for /room endpoint
Implemented a socket.io server ontop of the already existent express.js
server. Set the namespace to /room, so when the user hits the root (/)
endpoint no socket connection is established. Set up the js file linked
to the room.pug view template to have an insatce of Socket. (Took me a
lot of time...).