Implement named chat messages flow from index page
I have implememtned a full flow from the index page to the chat room.
Currently parsing name of user through url using GET verb, but I intend
on chaning for user login pages support Authentication. The chat room is
curently a single room and each message event is emitted to all client
socket connections, but this is also intended to change in the future. A
match making system of some sort would be ideal.