Add logic for a database and begin adding auth
Added a database to the app. I chose SQLite becasue it is serverless and
portable, it is also very effecient. This app does not require a super
optimizaed database, main focus is the functionality of the actual app.
Also chose the passport.js library for user authentication. I have began
implementing that logic using their docs/tutorials, usign google auth.