Add working implementation of passport-local auth
I was having errors and a hard time understadning auth in general by
using the passport-google strategy with oauth. Instead I added a classic
implementation of the passport-local strategy, where I store a user's
username (will probably change this to email) and password to my own
SQLite db and then can verify auth using the methods appended to the req
object by passport. This is not an end solution to auth for impostor as
obviously handling sensitive user data in-house is less secure and more
prone to spam attacks at the current level of implementation. Since this
app is not about auth, I also do not intend on making in-house local
auth as robust as possible. I am looking into other strategies such as
but not limited to JWT.