Add partial JWT auth
I began adding JWT auth to the app. I think I will stick with this auth
strategy for the near future becuase it is simple, convinient, and low
maintenance. I have not completed the implementation though. I also
think that I will remove support for google Oauth entirely. It is just
cluttering the codebase, and I have spend A LOT of time on auth already.
If this is reasonable to bring in the future, then I shall, but for now
JWT auth should do.