Full JWT support and removed Oauth support
I cleaned up the code base by removing any sessions auth related code.
This means I removed google Oauth support from the app, and the app is
now stateless. Completed the implementation of JWT auth using passport.
Happy to close the chapter on auth for the near future to focus more on
the functionality of the app. I do see how the current iteration of auth
and credential validation is not robust to spam and other attacks. But
this will do for now. I intend on implementing a more robust credential
system and perhaps sessions for a JWT blacklist. Or I might give Oauth
another try to delegate the auth work to the big tech companies who have
more robust systems.